+++ release_notes.md
@@ -15,6 +15,7 @@ PREESM Changelog
 * Move travis dist to xenial;
 * Fix PiSDF Exporter to read a PiMM instead of a scenario;
 * Fix Expression evaluator to properly handle several config input ports connected to the same parameter;
+* Copying the timing property during the PiMM2SRDAG conversion;
 
 ### Bug fix